ml_train/requirements.txt
click==7.1.2
dataclasses==0.7
fastapi==0.61.1
gunicorn==20.0.4
h11==0.11.0
httptools==0.1.1
pandas==1.1.3
psycopg2-binary==2.8.6
pydantic==1.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.12.2
uvloop==0.14.0
watchgod==0.6
websockets==8.1
